@@ -1,7 +1,6 @@
-﻿namespace UnitTestingSamples
+﻿namespace UnitTestingSamples;
+
+public class F1Addresses
 {
-    public class F1Addresses
-    {
-        public const string RacersUrl = "http://www.cninnovation.com/downloads/Racers.xml";
-    }
+    public const string RacersUrl = "http://www.cninnovation.com/downloads/Racers.xml";
 }
